backend/requirements.txt
# Core Flask
flask==3.0.0
flask-cors==4.0.0

# Git operations
GitPython==3.1.40

# Environment variables
python-dotenv==1.0.0

# AI - Anthropic Claude (updated version)
anthropic==0.39.0

# HTTP client (required by anthropic)
httpx==0.27.2

# Utility packages
requests==2.31.0

# Optional but recommended
Werkzeug==3.0.1